publisher/requirements.txt
azure-functions
azure-servicebus